+++ tests/data/test-file.txt
@@ -1,0 +1,1 @@
+This is a test file.
